faststack/pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "faststack"
version = "0.1.0"
authors = [
    { name="Gemini", email="gemini@google.com" },
]
description = "Ultra-fast JPG Viewer for RAW Stacking Selection"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "PySide6>=6.0,<7.0",
    "PyTurboJPEG>=1.8,<2.0",
    "numpy>=2.0,<3.0",
    "cachetools>=5.0,<6.0",
    "watchdog>=4.0,<5.0",
    "typer>=0.12,<1.0",
    "Pillow>=10.0,<11.0",
    "pytest>=8.0,<9.0",
]

[project.scripts]
faststack = "faststack.app:main"

[tool.setuptools]
packages = ["faststack"]
